Translator/requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
easygui==0.98.3
fonttools==4.53.1
idna==3.7
kiwisolver==1.4.5
matplotlib==3.9.1
numpy==2.0.1
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
PyAudio==0.2.14
pyparsing==3.1.2
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.16.0
SpeechRecognition==3.10.4
typing_extensions==4.12.2
urllib3==2.2.2
flask
flask_cors

